// Client/types/match.ts
import { Player, Position } from './player';

export interface TeamPlayer {
  player: Player;
  position: Position;
}

export interface Team {
  name: string;
  color: string;
  players: TeamPlayer[];
}

export interface Match {
  name: string;
  code: string;
  location: string;
  date: Date;
  time?: string;
  players_count: number;
  maxPlayers: number;
  isLive: boolean;
  formation: string;
  teams: Team[];
}